requirements.txt
# Data
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.24.0

# Database
# (sqlite3 is built-in)

# ML - LightGBM
lightgbm>=4.1.0
scikit-learn>=1.3.0
optuna>=3.4.0
shap>=0.43.0

# ML - CNN (required for hybrid model)
torch>=2.1.0

# Visualization + Labeling UI
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.8.0

# Config
pyyaml>=6.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0

# Testing
pytest>=7.4.0
